2013-03019: Fixed errors in tool chain build

+++ rpmbuild/TOOLS-LFS/builder.sh
@@ -7,16 +7,16 @@ _list=(binutils-pass-1 gcc-pass-1 linux-api-headers glibc binutils-pass-2 gcc-pa
 trap 'echo Toolchain build failed...;touch ${FAILURE};exit 1' ERR
 for i in ${_list[@]}; do
 	[ -f ${FAILURE} ] && (printf "Tool chain: Error exiting script \n";exit 0)
-	pushd "${i}"  > /dev/null 2>&1
+	#pushd "${i}"  > /dev/null 2>&1
 	if [ -e ${i}.done ]; then
 		echo "${i} --> Already Built"
 	else
-		unlink ${i}.log
+		[ -e ${i}.log ] && unlink ${i}.log
 		echo "Building---> ${i}"
 		( ./${i}.sh |& tee "${i}.log" ) || false
 		echo "Build ---> ${i} completed"
 		touch ${i}.done
 	fi
-	popd > /dev/null 2>&1
+	#popd > /dev/null 2>&1
 	rm -rf /tools/{,share}/{info,man,doc}
 done
